swclock init: Same file for both save and load
If no path is specified on the command-line when calling swclock, it
defaults to saving and loading the last shutdown time from a file named
shutdowntime.

Looking at the included init script for swclock, /etc/init.d/swclock, it
is calling swclock with no path defined when saving the time:
  swclock --save

But providing the path of @SBINDIR@/openrc-run when loading the time:
  swclock --warn @SBINDIR@/openrc-run

This results in the out of the box experience of /etc/init.d/swclock
being ineffective.