openrc: create lockfile for --user.
the pam module sets a variable that when in user mode, openrc should
pick up and either create/increment, or decrement. this is done so that
multiple session_open and session_close from pam doesn't randomly change
the user runlevel for the current user. only the first session open
starts the default runlevel, and only the last session close switches to
the none runlevel.

Signed-off-by: Anna (navi) Figueiredo Gomes <navi@vlhl.dev>