openrc-init: rework signal handling
Minimize the amount of work done in the signal handler by maintaining a
set of received signals and deferring the actual work to the main loop.

Treat SIGRTMIN+4 as a power-off signal for compatibility with
systemd-machined.

Adjust handle_shutdown to skip sending SIGTERM / SIGKILL if there are no
processes remaining.

Signed-off-by: Mike Gilbert <floppym@gentoo.org>