openrc-run: Allow multiplexed services to live outside init.d.
Use the realpath of the target to locate which system directory it's
in, using it as a base to load the config file. the service path is then
normalized from the symlink without dereferencing it.

Introduce a new variable for openrc-run, $RC_SYSCONFDIR, which points to
the respective system directory of the base service.

Signed-off-by: Anna (navi) Figueiredo Gomes <navi@vlhl.dev>