librc: Add initial support for dynamically multiplexed services.
Generate a new directory on the service dir, "multiplexed", which will
hold all the symlinks we generate at runtime, and a new librc function,
"rc_service_multiplex", which takes a basename or service path, and a
variant name, then generated a symlink to the base service from the
multiplex dir.

This allows multiplexing services at runtime, possibly using data
acquired at runtime, such as the name of an user or interface.