openrc-pam: Add pam module for openrc --user auto-start.
The module hooks on session start and end. If the user service is not
already multiplexed for a given user, it dynamically multiplexes it.

Login count is tracked under the "login" option for the user script, the
module checks it's value, and either starts the service hotplugged,
stops it, or skips.

Regardless, it also exports the XDG_RUNTIME_DIR variable to the user
session.

If the user service is already started and hotplugged, it does nothing.