init.d/swap: start after localmount
A swapfile can be placed on any partition, not just on root.
However because swap was started before localmount which is supposed to mount
any additional partitions swap would fail whenever the swapfile was located
on one of these additional partitions.
Making sure localmount ran before means the swapfile can be found and activated
successfully in all cases.