openrc-run: Allow symlinks to exist in different directories.
This is necessary for services in rc_svcdir() to be executed properly.
We then set RC_SVCNAME to applet, taken directly from argv[1], then we
call realpath on argv[1]. This allow us to execute `service` even after
chdir(), even if argv[1] is a relative path, and allows openrc-run.sh to
find conf.d since RC_SERVICE is derived from `service`, not applet.