supervise-daemon: use stpcpy and avoid quadratic loop
strcat has quadratic runtime behavior [0] due to having to walk
the dest string again and again. using stpcpy (POSIX function)
we can make it linear time instead.

admittedly, i don't think this will make a noticeable impact in
practice as argv strings won't get large enough for it to
matter. so this is mostly squashing a pet peeve of mine.

[0]: https://www.symas.com/post/the-sad-state-of-c-strings
Ref: https://github.com/OpenRC/openrc/issues/207