openrc-run: Fix loading of user configs for system-supplied user services
Without this patch, OpenRC loads user configuration files for
system-supplied services from `~/.config/rc/$SERVICE` instead of
`~/.config/rc/conf.d/$SERVICE`. Since the latter is documented in the
user-guide, and used for user services not supplied by the system, I
believe it to be the correct location.

Note that the `_usr_conf_d` variable is probably a bit inappropriately
named as the similarly named `_conf_d` points to the system conf.d
directory while `_usr_conf_d` points to the base rc configuration
directory. Therefore, while at it, rename `_usr_conf_d` to `_usr_conf`
to avoid this confusion in the future.

This is a fixup for ae9d743207f27481efe2c80593eea51cd8742ab2.