rc_pipe_command: implement it on top of rc_exec
since rc_exec uses an `execpipe` to detect exec failures inside
the child, this fixes the issue described in 792, but only for
the first layer of exec. there's a /bin/sh in the middle and we
won't be able to detect whether /bin/sh failed to exec the
command without inspecting the (unreliable) exit code.

Ref: https://github.com/OpenRC/openrc/issues/792