supervise-daemon: accept unit for respawn-{delay,period}
this is mainly because the current granularity is too coarse
going from no delay straight to 1 whole second. this allows
specified sub-second duration like "250ms", etc.

defaults to seconds for both backwards compatibility and
conventional reasons.

this also allows user to specify 0 as the duration (which is
documented as the default) because there's no reason not to.

Closes: https://github.com/OpenRC/openrc/pull/832