supervise-daemon: remove undocumented and unused -H option
this was added as a part of c1e582586, presumably as a
convenient way to specify a minutes timer. but it was never
documented, there's no init.d script variable for this and
finally, the number was never multiplied by 60, so it was
incorrect anyways.

given all that, just remove it. should be safe to do so.

if there's any actual demand for such convenience option, we can
change --healthcheck-timer to accept a duration similar to
--respawn-delay.