supervise-daemon: set default respawn-period to 12sec
currently respawn-period defaults to 0, which means if that a
services crashes 10 times (default respawn-max) in a period of 1
year, then supervise-daemon will just stop it completely. which
is very terrible default to say the least.

set it to 12sec by default instead. with the current default
respawn-delay of 0 and respawn-delay-step of 128ms it'd take
roughly ~6 seconds if a service instantly crashes 10 times
(current default respawn-max) in a row.