Reduce needless indentation
The parameter passed to -width is a raw string whose length is used to
determine the width of the left column.

The value being passed includes macros (which are not treated as such in
this context). The string is therefore longer than it needs to be, and
the leftmost column much wider than it needs to be.