supervisor=s6: rework support
 This commit modernizes the support for the supervisor=s6
backend. It reworks how the supervision tree is started and
how the services are defined.

 The s6-svscan service now runs its own catch-all logger,
logging to $RC_SVCDIR/s6-logs.

 A service defining supervisor=s6 now has an automatic "need"
dependency to s6-svscan. A service directory will be automatically
built from the information in the service file, under the
$RC_SVCDIR/s6-services repository. These service directories are
cached until the service file (or config) changes. So, services
don't need to provide their service directories themselves
anymore; they still can, but they'll need to do the s6-sv[un]link
thing themselves.

 The documentation has been updated.

 Anna, please add yourself to the AUTHORS file, because it is
ridiculous that I am now in it and you are not.