librc: clear dirfds when creating new svcdir
if for whatever reason, /run/openrc exists in the rootfs before first
boot, we'd still end up with an invalid fd. while users are not really
supposed to have anything in /run (under the tmpfs mount), should they
do, it's better to not fail

clearing the dirfds as we create svcdir also means we don't need a new
librc api function to clear it, for now at least.