openrc-run: remove timeout from start/stop
4d8a5c4 added the timeout logic from svc_wait into svc_exec, with the
goal of avoiding stalling the boot process should a service hang,
however we failed to mention the change in NEWS, and some critical
services like postgres rely on inifinite timeout for `stop()`.

this logic was also sharing the `timeout` keyword from `depend()`, which
is not very clean

let's revert the timeout logic, and properly re-introduce it later, with
effort put into making sure no issues happen.

Bug: https://github.com/OpenRC/openrc/issues/896